# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for Squeak Platform",
        "title": "Squeak API",
        "contact": {},
        "version": "1.0"
    },
    "host": "api.squeak.today",
    "basePath": "/",
    "paths": {
...
        "/billing": {
            "get": {
                "description": "Check Billing Account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "billing"
                ],
                "summary": "Check Billing Account",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.BillingAccountResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/billing/cancel-subscription-eop": {
            "post": {
                "description": "Cancel a Stripe individual subscription at the end of the period",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "billing"
                ],
                "summary": "Cancel a Stripe individual subscription at the end of the period",
                "parameters": [
                    {
                        "description": "Cancel subscription request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CancelIndividualSubscriptionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CancelIndividualSubscriptionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/billing/create-checkout-session": {
            "post": {
                "description": "Creates a checkout session and redirects to Stripe's payment page",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "billing"
                ],
                "summary": "Create a Stripe checkout session (individual)",
                "parameters": [
                    {
                        "description": "Create checkout session request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateIndividualCheckoutSessionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Redirect to Stripe Checkout",
                        "schema": {
                            "$ref": "#/definitions/models.CreateIndividualCheckoutSessionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/billing/usage": {
            "get": {
                "description": "Get Billing Account Usage, assumes free plan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "billing"
                ],
                "summary": "Get Billing Account Usage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Plan",
                        "name": "plan",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.BillingAccountUsageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/news": {
            "get": {
                "description": "Get news content by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "news"
                ],
                "summary": "Get news content",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Content ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetNewsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/news/query": {
            "get": {
                "description": "Get news content by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "news"
                ],
                "summary": "Get news content",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Language",
                        "name": "language",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "CEFR",
                        "name": "cefr",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Subject",
                        "name": "subject",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Page size",
                        "name": "pagesize",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.NewsItem"
                            }
                        }
                    }
                }
            }
        },
        "/profile": {
            "get": {
                "description": "Get the user's profile information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Get user profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetProfileResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.BillingAccountResponse": {
            "type": "object",
            "required": [
                "canceled",
                "expiration",
                "plan"
            ],
            "properties": {
                "canceled": {
                    "type": "boolean",
                    "example": false
                },
                "expiration": {
                    "type": "string",
                    "example": "2025-01-01T00:00:00Z"
                },
                "plan": {
                    "type": "string",
                    "example": "PRO"
                }
            }
        },
        "models.BillingAccountUsageResponse": {
            "type": "object",
            "required": [
                "max_natural_tts_usage",
                "max_premium_audiobooks_usage",
                "max_premium_stt_usage",
                "natural_tts_usage",
                "premium_audiobooks_usage",
                "premium_stt_usage"
            ],
            "properties": {
                "max_natural_tts_usage": {
                    "type": "integer",
                    "example": 100
                },
                "max_premium_audiobooks_usage": {
                    "type": "integer",
                    "example": 100
                },
                "max_premium_stt_usage": {
                    "type": "integer",
                    "example": 100
                },
                "natural_tts_usage": {
                    "type": "integer",
                    "example": 10
                },
                "premium_audiobooks_usage": {
                    "type": "integer",
                    "example": 10
                },
                "premium_stt_usage": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "models.CancelIndividualSubscriptionRequest": {
            "type": "object"
        },
        "models.CancelIndividualSubscriptionResponse": {
            "type": "object",
            "required": [
                "canceled_plan",
                "current_expiration",
                "success"
            ],
            "properties": {
                "canceled_plan": {
                    "type": "string",
                    "example": "PREMIUM"
                },
                "current_expiration": {
                    "type": "string",
                    "example": "2025-03-24T12:00:00Z"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.CreateIndividualCheckoutSessionRequest": {
            "type": "object"
        },
        "models.CreateIndividualCheckoutSessionResponse": {
            "type": "object",
            "required": [
                "redirect_url"
            ],
            "properties": {
                "redirect_url": {
                    "type": "string",
                    "example": "https://checkout.stripe.com/c/pay/123"
                }
            }
        },
        "models.ERROR_CODE": {
            "type": "string",
            "enum": [
                "PROFILE_NOT_FOUND",
                "NO_TRANSCRIPT",
                "AUTH_REQUIRED",
                "USAGE_LIMIT_REACHED",
                "USAGE_RESTRICTED"
            ],
            "x-enum-varnames": [
                "PROFILE_NOT_FOUND",
                "NO_TRANSCRIPT",
                "AUTH_REQUIRED",
                "USAGE_LIMIT_REACHED",
                "USAGE_RESTRICTED"
            ]
        },
        "models.ErrorResponse": {
            "type": "object",
            "required": [
                "error"
            ],
            "properties": {
                "code": {
                    "enum": [
                        "PROFILE_NOT_FOUND",
                        "NO_TRANSCRIPT",
                        "AUTH_REQUIRED",
                        "USAGE_LIMIT_REACHED",
                        "USAGE_RESTRICTED"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ERROR_CODE"
                        }
                    ],
                    "example": "PROFILE_NOT_FOUND"
                },
                "error": {
                    "type": "string",
                    "example": "Something went wrong"
                }
            }
        },
...
        "models.GetNewsResponse": {
            "type": "object",
            "required": [
                "cefr_level",
                "content",
                "content_type",
                "date_created",
                "dictionary",
                "language",
                "preview_text",
                "sources",
                "title",
                "topic"
            ],
            "properties": {
                "cefr_level": {
                    "type": "string",
                    "example": "B1"
                },
                "content": {
                    "type": "string",
                    "example": "Le contenu complet de l'article..."
                },
                "content_type": {
                    "type": "string",
                    "example": "News"
                },
                "date_created": {
                    "type": "string",
                    "example": "2024-02-26"
                },
                "dictionary": {
                    "$ref": "#/definitions/storage.Dictionary"
                },
                "language": {
                    "type": "string",
                    "example": "French"
                },
                "preview_text": {
                    "type": "string",
                    "example": "Un résumé des nouvelles musicales..."
                },
                "sources": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/storage.Source"
                    }
                },
                "title": {
                    "type": "string",
                    "example": "L'actualité musicale en bref"
                },
                "topic": {
                    "type": "string",
                    "example": "Music"
                }
            }
        },
        "models.GetProfileResponse": {
            "type": "object",
            "required": [
                "interested_topics",
                "learning_language",
                "skill_level",
                "username"
            ],
            "properties": {
                "daily_questions_goal": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 3
                },
                "interested_topics": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"NBA\"]"
                    ]
                },
                "learning_language": {
                    "type": "string",
                    "example": "French"
                },
                "skill_level": {
                    "type": "string",
                    "example": "B1"
                },
                "username": {
                    "type": "string",
                    "example": "connortbot"
                }
            }
        },
...
        "models.NewsItem": {
            "type": "object",
            "required": [
                "audiobook_tier",
                "cefr_level",
                "created_at",
                "date_created",
                "id",
                "language",
                "preview_text",
                "title",
                "topic"
            ],
            "properties": {
                "audiobook_tier": {
                    "type": "string",
                    "example": "NONE"
                },
                "cefr_level": {
                    "type": "string",
                    "example": "B1"
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-02-26T13:01:13.390612Z"
                },
                "date_created": {
                    "type": "string",
                    "example": "2024-02-26"
                },
                "id": {
                    "type": "string",
                    "example": "123"
                },
                "language": {
                    "type": "string",
                    "example": "French"
                },
                "preview_text": {
                    "type": "string",
                    "example": "Un résumé des nouvelles musicales..."
                },
                "title": {
                    "type": "string",
                    "example": "L'actualité musicale en bref"
                },
                "topic": {
                    "type": "string",
                    "example": "Music"
                }
            }
        },
...
        "storage.Dictionary": {
            "type": "object",
            "properties": {
                "translations": {
                    "type": "object",
                    "properties": {
                        "sentences": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        },
                        "words": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "storage.Source": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "score": {
                    "type": "number"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "JWT Authorization header using Bearer",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}